manifest: a new attempt at binary XML parsing/unparsing
this is some old work I have lying around. string-pool is taken from
the reapk trick: don't parse the XML, just rewrite strings in the
string-pool. this is simpler but now I want to parse/unparse the whole
of AndroidManifest.xml for "full" Android APK building support. or
something like that.