template ui: discard terminal output when things become slow
I don't know if this is a good idea, most terminals have a hardcoded
limit of like 500 000 lines. but it's not really what you care about:
you want to keep as much history as possible without things becoming
laggy. so let's try this and see how it goes.

instead of deleting the top half, we could keep it somewhere in case
it's important.